// src/interfaces.ts
export type onLoadT = 'login-required' | 'check-sso'
export type responseModeT = 'query' | 'fragment'
export type flowT = 'standard' | 'implicit' | 'hybrid'
export type adapterType = 'browser' | 'nativescript'

export abstract class InitOptions {
  /**
   * URL for the keycloak server
   */
  url: string
  /**
   * Realm to authenticate
   */
  realm: string
  /**
   * Client ID
   */
  clientId: string
  /**
   * Specifies an action to do on load. Supported values are 'login-required' or 'check-sso'.
   */
  onLoad?: onLoadT
  /**
   * Set an initial value for the token.
   */
  token?: string
  /**
   * Set an initial value for the refresh token.
   */
  refreshToken?: string
  /**
   * Set an initial value for the id token (only together with token or refreshToken).
   */
  idToken?: string
  /**
   * Set an initial value for skew between local time and Keycloak server in seconds (only together with token or refreshToken).
   */
  timeSkew?: number
  /**
   * Set to enable/disable monitoring login state (default is true).
   */
  checkLoginIframe?: boolean
  /**
   * Set the interval to check login state (default is 5 seconds).
   */
  checkLoginIframeInterval?: number
  /**
   * Set the OpenID Connect response mode send to Keycloak server at login request. Valid values are query or
   * fragment. Default value is fragment, which means that after successful authentication will Keycloak redirect
   * to javascript application with OpenID Connect parameters added in URL fragment. This is generally safer and
   * recommended over query.
   */
  responseMode?: responseModeT
  /**
   * Set the OpenID Connect flow. Valid values are standard, implicit or hybrid.
   */
  flow?: flowT

  adapter?: adapterType
  /**
   * Set the uri where application should be redirected after login
   */
  redirectUri?: string
}
export type promptT = 'none' | 'login'
export type actionT = 'register' | 'login'
export interface LoginOptions {
  /**
   * Specifies the uri to redirect to after login.
   */
  redirectUri?: string
  /**
   * By default the login screen is displayed if the user is not logged-in to Keycloak. To only authenticate to the
   * application if the user is already logged-in and not display the login page if the user is not logged-in, set this
   * option to none. To always require re-authentication and ignore SSO, set this option to login .
   */
  prompt?: promptT
  /**
   * Used just if user is already authenticated. Specifies maximum time since the authentication of user happened.
   * If user is already authenticated for longer time than maxAge, the SSO is ignored and he will need to
   * re-authenticate again.
   */
  maxAge?: number
  /**
   * Used to pre-fill the username/email field on the login form.
   */
  loginHint?: string
  /**
   * If value is 'register' then user is redirected to registration page, otherwise to login page.
   */
  action?: actionT
  /**
   * Specifies the desired locale for the UI.
   */
  locale?: string
  scope?: string
  idpHint?: string
}

export interface LogoutOptions {
  redirectUri: string
}